Model for the MerchStore
Why the addition of the model?
The addition of this model will help to manage the data and rules of the store, and it will provide a consistent way to access and manipulate that data.

Directory: backend\src\main\java\com\MerchStore\backend

What other things will your change affect?
N/A